Fix bug where content is not updated on navigate() call (#132)
* Refactor content logic into a hook

This hook detects changes in the initialContent for each composer
and updates the content as appropriate. This fixes the issue where
a command center navigation updates the title in the nav bar but
not the content.

* Don't return anything in useEffect()

The old code returns null if initialContent is empty. This causes
all sorts of React weirdness because useEffect() is either supposed
to not return anything, or return a cleanup function.