Keep disc modal visible while fetching a discussion
This was a hairy one to uncover. Once a discussion is created from
the backend, we update the modalDiscussionId. That triggers the
discussion query to fetch in the modal component. Since it's loading,
the component returns null, meaning the modal isn't visible anymore.

To fix, we ignore whether the query is loading and keep the modal
visible.