Discussion modal styling and bugfixes (#141)
* Discussion context styling

* Tweaks

* More tweaks

* Rename component and pass an explicit isModal state

Relying on modalRef is unreliable, because the ref may not be set
when it's read by a child component.

* Refactor AddReplyBox JSX

* Custom styles for message composer in modal

* Use padding instead of margin

* Quick code style tweak

* Fix showing message composer for general discussion in modal

* Keep disc modal visible while fetching a discussion

This was a hairy one to uncover. Once a discussion is created from
the backend, we update the modalDiscussionId. That triggers the
discussion query to fetch in the modal component. Since it's loading,
the component returns null, meaning the modal isn't visible anymore.

To fix, we ignore whether the query is loading and keep the modal
visible.

* Don't rely on state to set first message discussion id

It might not be set when the callback is triggered.

* Fix bug where messageCount is saved to wrong query in cache

* Don't write into cache if message exists

* Support arbitrary props for useQuery via usePaginatedResource

* Don't show discussion thread results until first message posted

* Prefetch discussion messages on inline discussion hover

* Refactor context generation logic into a hook

Helps with code readability and prepares us for more customization
in the hook. So that we don't confuse the logic in the component.

* Only truncate content for context when necessary

* More padding by default for inline context UI