Read/Unread and workspaces Pusher integration (#154)
* show unread indicator for sidebar resources

* saving progress

* wire up the badge_count mutation

* markAsRead for workspaces

* decrement badge_count when the notification is read

* minor code restructuring

* marking workspaceResourceAsRead changes

* show unread indicator for sidebar resources

* saving progress

* wire up the badge_count mutation

* markAsRead for workspaces

* decrement badge_count when the notification is read

* minor code restructuring

* marking workspaceResourceAsRead changes

* set documentId for only inline discussions

* markDocumentAsRead unconditionally

This allows marking the notification as read when a document is later added to the workspace. Also, this will help us know the last time the user read the document

* remove unused code

* workspaceResources query maynot be loaded

Also, fix the bug where workspace sidebar wasn't setting properly

* addressing PR review comments

* badgeCount should not be negative

This can happen is MarkAsRead is called multiple when the resource is Read. We can't use the readState because of useMountEffect so patching the localState to max of zero

Co-authored-by: recmend <arun@airseed.com>