Fix "cursor disappearing after autosave" regression
Instead of deselecting the editor in the cleanup function for the
useEffect() hook, do it within the component logic, but only when
we know the resource has changed.

To make this work we have to change the initialization of the
resourceId state. We initialize it to null, and only set it in the
useEffect() method. This ensures that when the resource ID has
changed, the new code will be able to detect a difference between
resourceId and initialResourceId.