Slightly better error handling on signup failure
Ideally, before we attempt the signup we check to see if the invite
code is valid. And if it's not valid, show a 404 page.

Bypassing that right now; we'll just display a generic error message
that doesn't make it seem like there's a bug elsewhere in the frontend
signup logic.